requirements.txt
streamlit==1.40.1
google-generativeai==0.7.2
Pillow==10.3.0
matplotlib==3.9.2